+++ src/icon/IconList.stories.tsx
@@ -1,32 +1,32 @@
-import { Meta, IconGallery, IconItem } from '@storybook/blocks';
-import Icon, { iconKey, IconCategory } from './Icon';
+import { Meta, IconGallery, IconItem } from "@storybook/blocks";
+import Icon, { iconKey, IconCategory } from "./Icon";
 
 export default {
-    title: 'Components/Icon/Icon List',
-    component: Icon,
-    tags: ['autodocs'],
+  title: "Icon/Icon List",
+  component: Icon,
+  tags: ["autodocs"],
 };
 
 export const Docs = () => <></>;
 
 Docs.parameters = {
-    docs: {
-        page: () => (
-            <>
-                <Meta title="Components/Icon/Icon List" />
-                {Object.entries(IconCategory).map(([category, icon]) => (
-                    <div key={category}>
-                        <div className="pb-7 text-black font-bold">{category}</div>
-                        <IconGallery>
-                            {Object.keys(icon).map((icon) => (
-                                <IconItem key={icon} name={icon as iconKey}>
-                                    <Icon icon={icon as iconKey} />
-                                </IconItem>
-                            ))}
-                        </IconGallery>
-                    </div>
-                ))}
-            </>
-        ),
-    },
+  docs: {
+    page: () => (
+      <>
+        <Meta title="Components/Icon/Icon List" />
+        {Object.entries(IconCategory).map(([category, icon]) => (
+          <div key={category}>
+            <div className="pb-7 text-black font-bold">{category}</div>
+            <IconGallery>
+              {Object.keys(icon).map(icon => (
+                <IconItem key={icon} name={icon as iconKey}>
+                  <Icon icon={icon as iconKey} />
+                </IconItem>
+              ))}
+            </IconGallery>
+          </div>
+        ))}
+      </>
+    ),
+  },
 };
